old_version/clinic_management_system - copy 1/requirements.txt
Flask
mysql-connector-python
pandas
secure-smtplib
email